💾 📦 Feat, Struct(Dashboard): 新增启动时显示公告的配置项, 移动公告管理器代码
